feat(tui): show weighted runner capacity
SASE_BEAD=[sase-z4.4][1]
SASE_PLAN=[202609/research_swarm_runners.md][2]
SASE_TYPE=stitch
SASE_AGENT=[bbugyi200.athena.sase-z4.4][3]

[1]: https://github.com/sase-org/sase--beads/blob/main/pages/sase-z4/sase-z4.4.md
[2]: https://github.com/sase-org/sase--plans/blob/main/202609/research_swarm_runners.md
[3]: https://github.com/sase-org/sase--agents/blob/main/agents/bbugyi200.athena.sase-z4.4/README.md

+++ src/sase/ace/tui/actions/agents/_loading_compute.py
@@ -168,7 +168,7 @@ def prepare_loaded_agents_apply_boundary(
     snapshot: PreparedApplySnapshot,
     *,
     merge_incomplete: bool = True,
-    effective_runner_limit: int | None = None,
+    effective_runner_limit: float | None = None,
 ) -> PreparedApplyBoundary:
     """Prepare pure post-load apply data from an explicit app-state snapshot."""
     from ...util.trace import tui_trace
